Fixing coutries triggers problems and changing neural network hidden layers
Changing the country area from integer to a float, and add the flag of the country to the response. Se 100 hidden layers. And add some intents

+++ language/countries.go
@@ -15,7 +15,7 @@ type Country struct {
 	SubRegion    string   `json:"subcontinent"`
 	Code         string   `json:"code"`
 	Borders      []string `json:"borders"`
-	Area         int      `json:"area"`
+	Area         float64  `json:"area"`
 	Currency     string   `json:"currency"`
 	Flag         string   `json:"flag"`
 }
